+++ packages/editor-core/src/index.ts
@@ -1,4 +1,5 @@
 export * from "./migrations";
+export * from "./projectMedia";
 export * from "./presets";
 export * from "./search";
 export * from "./types";